backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
pdfplumber
openai==0.28.1
google-generativeai
